requirements.txt
loguru==0.7.2
transformers==4.32.0
peft==0.7.0
torch>=2.0.1
deepspeed==0.12.4
torchvision
tiktoken==0.4.0
tensorboard==2.14.0
gradio==3.43.1
python-Levenshtein==0.12.2
numpy
ipywidgets
ipykernel
open-clip-torch==2.23.0
einops==0.6.1
